Move to stack instead of cabal
