+++ Arrow-interpreter.cabal
@@ -22,5 +22,5 @@ executable Arrow-interpreter
   other-modules:       Interp, Syntax
   other-extensions:    Arrows
   build-depends:       base >=4.12 && <4.13, containers >=0.6 && <0.7
-  -- hs-source-dirs:
+  hs-source-dirs:      src
   default-language:    Haskell2010
